+++ collector-exoplayer-example/src/main/java/com/bitmovin/analytics/exoplayer/example/MainActivity.kt
@@ -13,19 +13,16 @@ import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.enums.CDNProvider
 import com.bitmovin.analytics.example.shared.Sample
 import com.bitmovin.analytics.example.shared.Samples.DASH_DRM_WIDEVINE
+import com.bitmovin.analytics.example.shared.Samples.DASH_SINTEL
 import com.bitmovin.analytics.example.shared.Samples.HLS_REDBULL
 import com.bitmovin.analytics.exoplayer.ExoPlayerCollector
-import com.google.android.exoplayer2.C
+import com.bitmovin.analytics.exoplayer.example.databinding.ActivityMainBinding
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.MediaItem
 import com.google.android.exoplayer2.Player
 import com.google.android.exoplayer2.source.ConcatenatingMediaSource
+import com.google.android.exoplayer2.source.DefaultMediaSourceFactory
 import com.google.android.exoplayer2.source.MediaSource
-import com.google.android.exoplayer2.source.ProgressiveMediaSource
-import com.google.android.exoplayer2.source.dash.DashMediaSource
-import com.google.android.exoplayer2.source.hls.HlsMediaSource
-import com.google.android.exoplayer2.source.smoothstreaming.SsMediaSource
-import com.google.android.exoplayer2.ui.StyledPlayerView
 import com.google.android.exoplayer2.upstream.DataSource
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter
 import com.google.android.exoplayer2.upstream.DefaultDataSource
@@ -35,23 +32,27 @@ import com.google.android.exoplayer2.util.Util
 
 class MainActivity : AppCompatActivity(), DebugListener, Player.Listener {
 
+    private val viewBinding by lazy(LazyThreadSafetyMode.NONE) {
+        ActivityMainBinding.inflate(layoutInflater)
+    }
+
     private var player: ExoPlayer? = null
-    private var playerView: StyledPlayerView? = null
+
     private var dataSourceFactory: DataSource.Factory? = null
     private var bitmovinAnalytics: ExoPlayerCollector? = null
     private var bitmovinAnalyticsConfig: BitmovinAnalyticsConfig? = null
     private var eventLogView: TextView? = null
-    private val mediaSource: ConcatenatingMediaSource? = null
+    private var concatenatingMediaSource: ConcatenatingMediaSource? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(viewBinding.root)
         eventLogView = findViewById(R.id.eventLog)
         dataSourceFactory = DefaultDataSource.Factory(this, buildHttpDataSourceFactory())
 
         findViewById<Button>(R.id.release_button).setOnClickListener {
             if (player != null) {
-                player!!.release()
+                player?.release()
                 bitmovinAnalytics?.detachPlayer()
                 player = null
             }
@@ -60,51 +61,41 @@ class MainActivity : AppCompatActivity(), DebugListener, Player.Listener {
         findViewById<Button>(R.id.create_button).setOnClickListener { initializeExoPlayer() }
 
         findViewById<Button>(R.id.source_change_button).setOnClickListener {
-            bitmovinAnalytics!!.detachPlayer()
+            bitmovinAnalytics?.detachPlayer()
 
-            val mediaSource = buildMediaSource(DASH_DRM_WIDEVINE)
-            bitmovinAnalyticsConfig!!.videoId = "DRMVideo-id"
-            bitmovinAnalyticsConfig!!.title = "DRM Video Title"
+            val mediaSource = buildMediaItem(DASH_DRM_WIDEVINE)
+            bitmovinAnalyticsConfig?.videoId = "DRMVideo-id"
+            bitmovinAnalyticsConfig?.title = "DRM Video Title"
 
-            bitmovinAnalytics!!.attachPlayer(player!!)
-            player!!.setMediaSource(mediaSource)
+            bitmovinAnalytics?.attachPlayer(player!!)
+            player?.setMediaItem(mediaSource)
         }
 
         findViewById<Button>(R.id.set_custom_data).setOnClickListener {
             setCustomData()
         }
 
-        playerView = findViewById(R.id.a_main_exoplayer)
-
         initializeExoPlayer()
     }
 
-    private fun buildMediaSource(sample: Sample): MediaSource {
-        val uri = sample.uri
-        val type = Util.inferContentType(uri)
-        val builder = MediaItem.fromUri(uri).buildUpon()
+    private fun buildMediaItem(sample: Sample): MediaItem {
+        val mediaItemBuilder = MediaItem.Builder()
+            .setUri(sample.uri)
 
         val sampleDrmLicenseUri = sample.drmLicenseUri
         if (sample.drmScheme != null && sampleDrmLicenseUri != null) {
             val sampleDrmSchemeUUID = Util.getDrmUuid(sample.drmScheme!!)
             if (sampleDrmSchemeUUID != null) {
                 val drmConfiguration = MediaItem.DrmConfiguration.Builder(sampleDrmSchemeUUID)
-                        .setLicenseUri(sampleDrmLicenseUri)
-    //                    .setPlayClearContentWithoutKey(false)
-    //                    .setForceDefaultLicenseUri(true)
-                        .build()
-                builder.setDrmConfiguration(drmConfiguration)
+                    .setLicenseUri(sampleDrmLicenseUri)
+                    // .setPlayClearContentWithoutKey(false)
+                    // .setForceDefaultLicenseUri(true)
+                    .build()
+                mediaItemBuilder.setDrmConfiguration(drmConfiguration)
             }
         }
-        val mediaItem = builder.build()
-        val factory: MediaSource.Factory = when (type) {
-            C.TYPE_DASH -> DashMediaSource.Factory(dataSourceFactory!!)
-            C.TYPE_SS -> SsMediaSource.Factory(dataSourceFactory!!)
-            C.TYPE_HLS -> HlsMediaSource.Factory(dataSourceFactory!!)
-            C.TYPE_OTHER -> ProgressiveMediaSource.Factory(dataSourceFactory!!)
-            else -> throw IllegalStateException("Unsupported type: $type")
-        }
-        return factory.createMediaSource(mediaItem)
+
+        return mediaItemBuilder.build()
     }
 
     private fun buildHttpDataSourceFactory(): HttpDataSource.Factory {
@@ -112,61 +103,66 @@ class MainActivity : AppCompatActivity(), DebugListener, Player.Listener {
             .setUserAgent(Util.getUserAgent(this, getString(R.string.app_name)))
     }
 
-    private fun buildConcatenatingMediaSource(): ConcatenatingMediaSource {
-        val concatenatingMediaSource = ConcatenatingMediaSource()
-        var mediaSource: MediaSource = buildMediaSource(HLS_REDBULL)
-        concatenatingMediaSource.addMediaSource(mediaSource)
-        mediaSource = buildMediaSource(HLS_REDBULL)
-        concatenatingMediaSource.addMediaSource(mediaSource)
-        return concatenatingMediaSource
+    private fun buildConcatenatingMediaSource() {
+       val mediaSourceFactory = DefaultMediaSourceFactory(applicationContext)
+        this.concatenatingMediaSource = ConcatenatingMediaSource()
+        var mediaSource: MediaSource = mediaSourceFactory.createMediaSource(buildMediaItem(HLS_REDBULL))
+        concatenatingMediaSource?.addMediaSource(mediaSource)
+        mediaSource = mediaSourceFactory.createMediaSource(buildMediaItem(DASH_SINTEL))
+        concatenatingMediaSource?.addMediaSource(mediaSource)
     }
 
     private fun initializeExoPlayer() {
         if (player == null) {
             val bandwidthMeter = DefaultBandwidthMeter.Builder(this).build()
 
-            val exoBuilder = ExoPlayer.Builder(this)
-            exoBuilder.setBandwidthMeter(bandwidthMeter)
+            player = ExoPlayer.Builder(this)
+                .setBandwidthMeter(bandwidthMeter)
+                .build()
+                .also { exoPlayer ->
+                    viewBinding.aMainExoplayer.player = exoPlayer
+                    exoPlayer.addListener(this)
+                    // Step 1: Create your analytics config object
+                    bitmovinAnalyticsConfig = createBitmovinAnalyticsConfig()
+                    eventLogView?.text = ""
 
-            player = exoBuilder.build()
-            player!!.addListener(this)
+                    // Step 2: Create Analytics Collector
+                    bitmovinAnalytics = ExoPlayerCollector(
+                        bitmovinAnalyticsConfig!!,
+                        applicationContext
+                    )
 
-            // Step 1: Create your analytics config object
-            bitmovinAnalyticsConfig = createBitmovinAnalyticsConfig()
-            eventLogView!!.text = ""
+                    bitmovinAnalytics?.addDebugListener(this)
+                    bitmovinAnalytics = bitmovinAnalytics
 
-            // Step 2: Create Analytics Collector
-            bitmovinAnalytics = ExoPlayerCollector(
-                bitmovinAnalyticsConfig!!,
-                applicationContext
-            )
-            bitmovinAnalytics!!.addDebugListener(this)
-            bitmovinAnalytics = bitmovinAnalytics
+                    // Step 3: Attach ExoPlayer
+                    bitmovinAnalytics?.attachPlayer(exoPlayer)
 
-            // Step 3: Attach ExoPlayer
-            bitmovinAnalytics?.attachPlayer(player!!)
+                    // Step 4: Create, prepare, and play media source
+                    // val mediaItem = buildMediaItem(HLS_REDBULL)
+                    // exoPlayer.setMediaItem(mediaItem)
 
-            // Step 4: Create, prepare, and play media source
-            playerView!!.player = player
+                    this.buildConcatenatingMediaSource()
+                    exoPlayer.setMediaSource(this.concatenatingMediaSource!!)
 
-            val mediaSource = buildMediaSource(HLS_REDBULL)
-            player!!.setMediaSource(mediaSource)
+                    // autoplay
+                    exoPlayer.playWhenReady = false
 
-//            val concatenatingMediaSource: ConcatenatingMediaSource = buildConcatenatingMediaSource()
-//            player!!.setMediaSource(concatenatingMediaSource)
-
-            player!!.playWhenReady = false
-            // without prepare() it will not start autoplay
-            // prepare also starts preloading data before user clicks play
-            player!!.prepare()
+                    // without prepare() it will not start autoplay
+                    // prepare also starts preloading data before user clicks play
+                    exoPlayer.prepare()
+                }
         }
     }
 
     private fun setCustomData() {
-        val customData = bitmovinAnalytics!!.customData
-        customData.customData1 = "custom_data_1_changed"
-        customData.customData2 = "custom_data_2_changed"
-        bitmovinAnalytics!!.customData = customData
+        val customData = bitmovinAnalytics?.customData
+
+        if (customData != null) {
+            customData.customData1 = "custom_data_1_changed"
+            customData.customData2 = "custom_data_2_changed"
+            bitmovinAnalytics?.customData = customData
+        }
     }
 
     private fun createBitmovinAnalyticsConfig(): BitmovinAnalyticsConfig {
@@ -194,20 +190,22 @@ class MainActivity : AppCompatActivity(), DebugListener, Player.Listener {
 
     private var oldIndex = 0
     override fun onPositionDiscontinuity(reason: Int) {
-        val sourceIndex = player!!.currentMediaItemIndex
+        val sourceIndex = player?.currentMediaItemIndex
         if (sourceIndex != oldIndex) {
             if (oldIndex >= 0) {
-                mediaSource!!.removeMediaSource(
+                concatenatingMediaSource?.removeMediaSource(
                     oldIndex,
                     Handler(Looper.getMainLooper())
                 ) {
-                    println("Mainactivity isPlaying: " + player!!.isPlaying)
-                    println("Mainactivity playbackState: " + player!!.playbackState)
-                    println("Mainactivity playWhenReady: " + player!!.playWhenReady)
-                    bitmovinAnalytics!!.attachPlayer(player!!)
+                    println("Mainactivity isPlaying: " + player?.isPlaying)
+                    println("Mainactivity playbackState: " + player?.playbackState)
+                    println("Mainactivity playWhenReady: " + player?.playWhenReady)
+                    bitmovinAnalytics?.attachPlayer(player!!)
                 }
             }
-            oldIndex = sourceIndex
+            if (sourceIndex != null) {
+                oldIndex = sourceIndex
+            }
         }
     }